requirements.txt
arxiv
langchain
langchain_community
streamlit
pandas
tqdm
sentence-transformers
faiss-cpu
numpy
python-dotenv
ollama
scikit-learn
plotly
google-generativeai
pypdf